Split Sun atmosphere data owner
