+++ scripts/quality-baseline.json
@@ -7,6 +7,6 @@
   "viewRuntimeBridgeLookups": 0,
   "labStateAppFiles": 0,
   "labStateTestFiles": 0,
-  "largeJsFilesOver800Lines": 7,
+  "largeJsFilesOver800Lines": 6,
   "maxJsFileLines": 1168
 }
