+++ tests/test-correctness-phase2.js
@@ -198,6 +198,7 @@ const pwaAppShellAssets = [
   '/js/sun-spectrum-actions.js',
   '/js/sun-spectrum-device.js',
   '/js/sun-spectrum.js',
+  '/js/sun-uvdata-atmosphere.js',
   '/js/sun-uvdata-config.js',
   '/js/sun-uvdata.js',
   '/js/chat-window-bindings.js',
