+++ tsconfig.checkjs.json
@@ -301,6 +301,7 @@
     "js/sun-spectrum-actions.js",
     "js/sun-spectrum-device.js",
     "js/sun-spectrum.js",
+    "js/sun-uvdata-atmosphere.js",
     "js/sun-uvdata-config.js",
     "js/sun-uvdata.js",
     "js/sun.js",
